requirements.txt
pydantic
tenacity
pytest
fastapi
uvicorn
supabase
python-dotenv
openai
anthropic
python-multipart
